checkSonify {playitbyr}	R Documentation

Check if a sonify object can be rendered

Description

This function is called by print.sonify before rendering to check if the object can be rendered; it can also be called directly by the user for diagnostic purposes.

Usage

checkSonify(x)

Arguments

x 	 A sonify object.

See Also

sonify, sonaes, sonscaling all give more details on how to correctly form a sonify object.

[Package playitbyr version 0.1-1 Index Web site]