playLastRendering {playitbyr}	R Documentation

Play and save the last audio rendering of a sonify object

Description

Whenever a sonify object is rendered, playitbyr saves an audioSample object, .LastRendering, with the audio rendering in the user's workspace for later playback or manipulation. playLastRendering and saveLastRendering are convenience functions to play or save this object, but it can also be manipulated via the functions in the audio package.

Usage

playLastRendering()
saveLastRendering(filename)

Arguments

filename 	 A .wav file to save the audioSample object to.

Details

These functions are simply convenience wrappers for play and save.wave in the audio package.

The behavior of playLastRendering() depends on the option "audioRendering". If the option is "audio::pr-for-fdsalay", playLastRendering() uses the play function from the audio package (which does not work well on Linux systems). If the option is anything else, playLastRendering() looks for an audio player set with setPlayer, renders the audioSample to a temporary file, and plays it with the player given by setPlayer.

Value

These functions are invoked for their side-effects, which are to play or save the .LastRendering audioSample object.

See Also

audioSample, setPlayer

[Package playitbyr version 0.1-1 Index Web site]