render {playitbyr}	R Documentation

Render sonify objects to audioSample

Usage

render(x)
playAudioRendering(audioSamp)

Arguments

x 	 A sonify object to be rendered.

audioSamp 	 An audioSample object, already rendered and ready to be played or saved to file.

Details

print.sonify calls both these functions; these are provided in case you want to manipulate the object further or save it to file.

Value

render.audio returns an audioSample object (from the audio package) of the sonification.

playAudioRendering is called for its side effect, to produce the sound. It is just a wrapper for play on Windows and OS X; on Linux systems it creates a temporary file and plays it with an external *.wav file player (see setPlayer).

See Also

save.wave for saving audioSample objects to file; setPlayer for getting and setting the method of actually playing the rendering; and sonify for general details.

Examples

x <- sonify(iris) + sonaes(Petal.Length, Petal.Width)
out <- render(x)
## Not run: 
playAudioRendering(out)
## This is the same as just printing the object:
x

## But you can also save it to file using the audio package:
save.wave(out, "myfile.wav")

## End(Not run)

[Package playitbyr version 0.1-1 Index Web site]