scaleShortcuts {playitbyr}	R Documentation

Shortcut functions to set linear scales for sonify objects

Description

These shortcut functions (with the notation scale_[data parameter]_[scaling function]) are intended to be added onto a sonify object to specify how [data parameter] from the data (a column of the data) is to be mapped to sonic parameters.

Usage

scale_time_linear(min, max)
scale_pitch_linear(min, max)
scale_dur_linear(min, max)
scale_vol_linear(min, max)
scale_pan_linear(min, max)
scale_tempo_linear(min, max)

Arguments

min 	 The desired minimum value, a numeric of length 1

max 	 The desired maximum value, a numeric of length 1

Details

The syntax of specifying a full sonify object can be rather verbose. Inspired by the ggplot2 package, playitbyr provides shortcut functions to make the process easier and more intuitive.

These functions are intended to be added onto a pre-existing sonify object, as shown in the examples.

Value

A sonscaling object, to be used in sonify or added onto a sonify object.

See Also

sonscaling for the full syntax and range of possibilities for specifying scaling of sonify objects; linear.scale for the pre-defined linear scaling function which this incorporates; +.sonify for the addition onto sonify objects.

Examples

## A verbose way of specifying a sonify object,
## without using the shortcuts
x <- sonify(data=iris, 
     sonaes(Petal.Width, Petal.Length, dur=5, vol=0.75),
     scales = sonscaling(
       time = list(min=0, max=3,
        scaling.function = linear.scale),
       pitch = list(min=3, max=13,
        scaling.function = linear.scale)))
summary(x)
## Not run: x

## An equivalent and much more compact way, with
## the shortcut functions
y <- sonify(iris, sonaes(Petal.Width, Petal.Length, dur=5, vol=0.75))
y <- y + scale_time_linear(0, 3)
y <- y + scale_pitch_linear(3, 13)
summary(y)
## Not run: y

## You can also change a mapping incrementally.
## New parameters overwrite old ones, so now
## the sonification is stretched to 12 seconds
## long instead of 3, and the pitch is mapped to
## a much narrower range, the octave below tuning A (440 Hz)
y <- y + scale_time_linear(0, 12) + scale_pitch_linear(7, 8)
summary(y)
## Not run: y

[Package playitbyr version 0.1-1 Index Web site]