sonaes {playitbyr}	R Documentation

Create mappings and parameters for sonification objects

Description

Creates a sonaes object, which is a list containing the desired mappings of data columns or constants onto each sound parameter. This is intended be added to a sonify object (or included in its construction).

Usage

sonaes(time=0, pitch=8, dur=2, vol=1, pan=0.5, tempo=NULL,
       timbre="sine")

Arguments

time 	 a data.frame column name corresponding to one of the columns of the data argument of sonify, or the desired start time of all events (in seconds). Either time or tempo must be specified, but not both, for a sonify object to be rendered. For any of these arguments, quotes are optional.

pitch 	 a data.frame column name or the desired pitch of all events, represented such that 8 represents middle C, 9 represents the octave above, etc. (This is Csound's ‘oct’ notation.)

dur 	 a data.frame column name, or the relative desired duration of all events in beats, where 1 beat equals the length of time for one event if all events were equal length.

vol 	 a data.frame column name, or the desired volume of all events as a number between 0, silence, and 1, the maximum possible amplitude

pan 	 a data.frame column name, or the desired balance between left and right stereo channels, where 0 is all the left channel, 0.5 balanced, and 1 all the right.

tempo 	 a data.frame column name, or the desired tempo of all events (in beats per minute)

timbre 	 The desired timbre of the sound. Currently only the default value "sine", for a simple sine wave, is supported.

Details

Setting sonic parameters to data.frame columns or constant values is the heart of making the sonification happen and making it possible to render.

All fields, except for one of time or tempo, must be assinged in the sonify before rendering to sound. sonaes does not check to make sure that all mappings are filled in. since the user can add on more mappings (using +.sonify) interectively. However, this is checked by checkSonify when the object is to be rendered.

Also, every item that is mapped to a data.frame column in a sonify object must also have a scale associated with it before rendering; see sonscaling.

Value

A sonaes object, used in or added to a sonify object.

See Also

sonify, sonscaling, octToFreq . Also, see aes from the ggplot2 package, which inspired this function.

Examples

## Maps Petal.Width onto tempo,
## Sepal.Width onto pitch,
## and sets scalings
## (not terribly informative, since the order is
## arbitrary)
x <- sonify(iris, sonaes(tempo=Petal.Width,
            pitch=Sepal.Width))
summary(x)
## Not run: x ## Quite long!

## Use a different tempo scaling than default
## to hear more of the data faster
## Not run: x + scale_tempo_linear(300, 1000)

## Map Petal.Width onto 'time' instead
y <- x + sonaes(time=Petal.Width)
summary(y)
## Not run: y

## Seems clipped, so increase duration
## Not run: y + sonaes(dur=4)

[Package playitbyr version 0.1-1 Index Web site]