sonifyclass {playitbyr}	R Documentation

print, summary and addition methods for sonify objects

Description

Methods to print and alter sonify objects. Printing a sonify object renders it to sound (analogously to how printing ggplot objects renders them to screen in the ggplot2 package).

Usage

## S3 method for class 'sonify'
print(x, ...)
## S3 method for class 'sonify'
summary(object, ...)
## S3 method for class 'sonify'
x + y
x %+% y

Arguments

x, object 	 A sonify object

y 	 For +, a sonscaling object generated by scaling() or a sonaes object generated by sonaes() to add to x.
For %+%, a data.frame to add to x.

... 	 Ignored.

Details

The print method for sonify objects not only renders them to sound, it also creates an audioSample object (from the audio package), .LastRendering, in the user's workspace. .LastRendering is the output of the sonification and can be replayed or saved to a .wav file without regenerating the sonification from scratch with the convenience functions playLastRendering and saveLastRendering.

Value

+ and %+% both return a sonify object with the relevant y value added into the object.

print.sonify is called for its side-effect, which is to actually render the object to a sound.

summary.sonify prints a brief summary of the name of the dataset to be sonified, and the specified mappings and scalings of sonic values to data parameters.

See Also

sonify for the creation of these objects.

[Package playitbyr version 0.1-1 Index Web site]